Refactor: Move OrderToggleButton to /admin/order
